address/pyproject.toml
[project]
name = "address-service"
version = "0.1.0"
description = "Location-based address service with geocoding"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.20.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "geoalchemy2>=0.14.3",
    "alembic>=1.13.1",
    "redis>=5.0.1",
    "httpx>=0.26.0",
    "googlemaps>=4.10.0",
    "geopy>=2.4.1",
    "geopandas>=0.14.2",
    "shapely>=2.0.2",
    "pyproj>=3.6.1",
    "pandas>=2.1.4",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "tenacity>=8.2.3",
    "prometheus-client>=0.19.0",
    "structlog>=24.1.0",
    "h3>=3.7.6",
    "folium>=0.15.1",
    "rtree>=1.1.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "black>=23.12.1",
    "ruff>=0.1.11",
    "mypy>=1.8.0",
    "ipython>=8.19.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "black>=23.12.1",
    "ruff>=0.1.11",
]